Fixed bug with TicksExpression and many changes

+++ engine/src/engine/simulation/world/expression/auxiliary/RandomExpression.java
@@ -33,7 +33,7 @@ public class RandomExpression extends AbstractExpression {
 
     @Override
     public ValueType getValueType() {
-        return ValueType.DECIMAL;
+        return ValueType.FLOAT;
     }
 
     @Override
